removed seed checking to allow for different seed types
